Astronaut + Cars
Added a wall of cars and redesigned the ground floor of the level. Now, replaced the alien with an astronaut character, since the collisions and rigidbody had too many issues, and edited the code in the given astronaut movement code to fit the project and be written in my own style.